+++ fltk/src/input.rs
@@ -56,6 +56,22 @@ pub struct FileInput {
     _tracker: *mut fltk_sys::fl::Fl_Widget_Tracker,
 }
 
+impl FileInput {
+    pub fn set_down_frame(&mut self, f: FrameType) {
+        assert!(!self.was_deleted());
+        unsafe {
+            Fl_File_Input_set_down_box(self._inner, f as i32)
+        }
+    }
+    
+    pub fn down_frame(&self) -> FrameType {
+        assert!(!self.was_deleted());
+        unsafe {
+            mem::transmute(Fl_File_Input_down_box(self._inner))
+        }
+    }
+}
+
 /// Creates a secret input widget
 #[derive(WidgetBase, WidgetExt, InputExt, Debug)]
 pub struct SecretInput {